pyproject.toml
[tool.poetry]
name = "notification-service"
version = "0.1.0"
description = ""
authors = ["Almaz Durusaliev"]
readme = "README.md"
packages = [{include = "notification_service"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.1"
djangorestframework = "^3.14.0"
psycopg2-binary = "^2.9.6"
celery = "^5.2.7"
redis = "^4.5.5"
django-celery-beat = "^2.5.0"
drf-yasg = "^1.21.5"
python-decouple = "^3.8"
twilio = "^8.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
